# tsoa/swagger.json
{
	"components": {
...
		"schemas": {
...
			"RestaurantSummary": {
				"properties": {
					"price_average": {
						"type": "number",
						"format": "double"
					},
					"rating_count": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"price_average",
					"rating_count",
					"rating"
				],
				"type": "object"
			},
...
			"Pick_Reservation.number_of_person-or-date_of_reservation_": {
				"properties": {
					"number_of_person": {
						"type": "number",
						"format": "double"
					},
					"date_of_reservation": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"number_of_person",
					"date_of_reservation"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"ReservationDetailOutput": {
				"$ref": "#/components/schemas/Pick_Reservation.number_of_person-or-date_of_reservation_"
			},
			"RestaurantCardOutput_ReservationDetailOutput_": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"nullable": true
					},
					"brand_name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"header_url": {
						"type": "string",
						"nullable": true
					},
					"restaurant_type": {
						"type": "string"
					},
					"location": {},
					"summary": {
						"$ref": "#/components/schemas/RestaurantSummary"
					},
					"detail": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ReservationDetailOutput"
							}
						],
						"nullable": true
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"total_bookings": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"status",
					"brand_name",
					"address",
					"city",
					"header_url",
					"restaurant_type",
					"location",
					"summary",
					"detail",
					"created_at",
					"total_bookings"
				],
				"type": "object",
				"additionalProperties": false
			},
			"%24Result.DefaultSelection_Prisma.%24ReservationPayload_": {
				"properties": {
					"restaurant_id": {
						"type": "string"
					},
					"user_id": {
						"type": "string"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"type": "string"
					},
					"date_of_reservation": {
						"type": "string",
						"format": "date-time"
					},
					"number_of_person": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"restaurant_id",
					"user_id",
					"updated_at",
					"created_at",
					"status",
					"date_of_reservation",
					"number_of_person",
					"id"
				],
				"type": "object"
			},
			"Reservation": {
				"$ref": "#/components/schemas/%24Result.DefaultSelection_Prisma.%24ReservationPayload_",
				"description": "Model Reservation"
			},
			"Pick_UserRecord.firstname-or-lastname-or-avatar_url_": {
				"properties": {
					"firstname": {
						"type": "string"
					},
					"lastname": {
						"type": "string",
						"nullable": true
					},
					"avatar_url": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"firstname",
					"lastname",
					"avatar_url"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"UserInfo": {
				"$ref": "#/components/schemas/Pick_UserRecord.firstname-or-lastname-or-avatar_url_"
			},
			"WithUserInfo_Reservation_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Reservation"
					},
					{
						"properties": {
							"user": {
								"$ref": "#/components/schemas/UserInfo"
							}
						},
						"required": [
							"user"
						],
						"type": "object"
					}
				]
			},
			"ReservationOutput": {
				"$ref": "#/components/schemas/WithUserInfo_Reservation_"
			},
			"ReservationCreateInput": {
				"$ref": "#/components/schemas/ReservationDetailOutput"
			},
			"Pick_Reservation.status_": {
				"properties": {
					"status": {
						"type": "string"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"ReservationUpdateInput": {
				"$ref": "#/components/schemas/Pick_Reservation.status_"
			},
			"JsonObject": {
				"properties": {},
				"type": "object",
				"description": "From https://github.com/sindresorhus/type-fest/\nMatches a JSON object.\nThis type can be useful to enforce some input to be JSON-compatible or as a super-type to be extended from."
			},
			"JsonArray": {
				"description": "From https://github.com/sindresorhus/type-fest/\nMatches a JSON array.",
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"Prisma.JsonValue": {
				"anyOf": [
					{
						"type": "string"
					},
					{
						"type": "number",
						"format": "double"
					},
					{
						"type": "boolean"
					},
					{
						"$ref": "#/components/schemas/JsonObject"
					},
					{
						"$ref": "#/components/schemas/JsonArray"
					}
				],
				"nullable": true,
				"description": "From https://github.com/sindresorhus/type-fest/\nMatches any valid JSON value."
			},
			"RestaurantRecord": {
				"properties": {
					"phone": {
						"type": "string",
						"nullable": true
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"address": {
						"type": "string",
						"nullable": true
					},
					"country": {
						"type": "string",
						"nullable": true
					},
					"city": {
						"type": "string",
						"nullable": true
					},
					"restaurant_type": {
						"type": "string",
						"nullable": true
					},
					"location": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Prisma.JsonValue"
							}
						],
						"nullable": true
					},
					"name": {
						"type": "string"
					},
					"header_url": {
						"type": "string",
						"nullable": true
					},
					"brand_name": {
						"type": "string",
						"nullable": true
					},
					"price_average": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"rating_count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"phone",
					"description",
					"address",
					"country",
					"city",
					"restaurant_type",
					"location",
					"name",
					"header_url",
					"brand_name",
					"price_average",
					"rating",
					"rating_count"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_RestaurantRecord.-or-name-or-brand_name-or-address-or-price_average-or-header_url-or-rating-or-rating_count-or-city-or-country-or-restaurant_type_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"address": {
						"type": "string",
						"nullable": true
					},
					"country": {
						"type": "string",
						"nullable": true
					},
					"city": {
						"type": "string",
						"nullable": true
					},
					"restaurant_type": {
						"type": "string",
						"nullable": true
					},
					"header_url": {
						"type": "string",
						"nullable": true
					},
					"brand_name": {
						"type": "string",
						"nullable": true
					},
					"price_average": {
						"type": "number",
						"format": "double"
					},
					"rating_count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"rating",
					"address",
					"country",
					"city",
					"restaurant_type",
					"header_url",
					"brand_name",
					"price_average",
					"rating_count"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"RestaurantCardRecord": {
				"$ref": "#/components/schemas/Pick_RestaurantRecord.-or-name-or-brand_name-or-address-or-price_average-or-header_url-or-rating-or-rating_count-or-city-or-country-or-restaurant_type_"
			},
			"LandingRestaurantInfo": {
				"properties": {},
				"additionalProperties": {
					"items": {
						"$ref": "#/components/schemas/RestaurantCardRecord"
					},
					"type": "array"
				},
				"type": "object"
			},
			"RestaurantCardOutput_unknown_": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"nullable": true
					},
					"brand_name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"header_url": {
						"type": "string",
						"nullable": true
					},
					"restaurant_type": {
						"type": "string"
					},
					"location": {},
					"summary": {
						"$ref": "#/components/schemas/RestaurantSummary"
					},
					"detail": {
						"nullable": true
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"total_bookings": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"status",
					"brand_name",
					"address",
					"city",
					"header_url",
					"restaurant_type",
					"location",
					"summary",
					"detail",
					"created_at",
					"total_bookings"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResultsSummary_RestaurantCardOutput_unknown__": {
				"properties": {
					"results": {
						"items": {
							"$ref": "#/components/schemas/RestaurantCardOutput_unknown_"
						},
						"type": "array"
					},
					"options": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"results",
					"options",
					"count"
				],
				"type": "object"
			},
			"SortRestaurantBy": {
				"type": "string",
				"enum": [
					"rating",
					"visits",
					"new"
				]
			},
			"Pick_UserOutput.email-or-password-or-avatar_url_": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"avatar_url": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"password",
					"email",
					"avatar_url"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"AdministratorInput": {
				"$ref": "#/components/schemas/Pick_UserOutput.email-or-password-or-avatar_url_"
			},
			"Pick_RestaurantProps.name_": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"RestaurantInput": {
				"$ref": "#/components/schemas/Pick_RestaurantProps.name_"
			},
			"Pick_InformationProps.phone-or-address-or-country-or-city-or-restaurant_type-or-postal_code_": {
				"properties": {
					"phone": {
						"type": "string",
						"nullable": true
					},
					"address": {
						"type": "string",
						"nullable": true
					},
					"country": {
						"type": "string",
						"nullable": true
					},
					"city": {
						"type": "string",
						"nullable": true
					},
					"restaurant_type": {
						"type": "string",
						"nullable": true
					},
					"postal_code": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"phone",
					"address",
					"country",
					"city",
					"restaurant_type",
					"postal_code"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"InformationInput": {
				"$ref": "#/components/schemas/Pick_InformationProps.phone-or-address-or-country-or-city-or-restaurant_type-or-postal_code_"
			},
			"RestaurantCreateInput": {
				"allOf": [
					{
						"$ref": "#/components/schemas/AdministratorInput"
					},
					{
						"$ref": "#/components/schemas/RestaurantInput"
					},
					{
						"$ref": "#/components/schemas/InformationInput"
					},
					{
						"properties": {
							"brand_name": {
								"type": "string",
								"nullable": true
							}
						},
						"required": [
							"brand_name"
						],
						"type": "object"
					}
				]
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "resdy-back",
		"version": "1.0.0",
		"description": "Rest API with Nodejs and Prisma",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Carlos Velázquez"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/reservation/myReservations": {
			"get": {
				"operationId": "GetMyReservations",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RestaurantCardOutput_ReservationDetailOutput_"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Reservation service"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "city",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "start_date",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "end_date",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/reservation/{restaurant_name}": {
			"get": {
				"operationId": "GetReservations",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ReservationOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Reservation service"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "restaurant_name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "PostReservation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Reservation"
								}
							}
						}
					}
				},
				"tags": [
					"Reservation service"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "restaurant_name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReservationCreateInput"
							}
						}
					}
				}
			}
		},
		"/reservation/{reservation_id}": {
			"put": {
				"operationId": "PutReservation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Reservation"
								}
							}
						}
					}
				},
				"tags": [
					"Reservation service"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "reservation_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReservationUpdateInput"
							}
						}
					}
				}
			}
		},
		"/restaurant": {
			"get": {
				"operationId": "GetRestaurant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RestaurantRecord"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Restaurant service"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "city",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "restaurant_type",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "country",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateRestaurant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {
											"type": "string"
										}
									},
									"required": [
										"token"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Restaurant service"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RestaurantCreateInput"
							}
						}
					}
				}
			}
		},
		"/restaurant/landing": {
			"get": {
				"operationId": "GetLandingRestaurant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LandingRestaurantInfo"
								}
							}
						}
					}
				},
				"tags": [
					"Restaurant service"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "city",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "country",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/restaurant/discover": {
			"get": {
				"operationId": "GetDiscoverRestaurant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResultsSummary_RestaurantCardOutput_unknown__"
								}
							}
						}
					}
				},
				"tags": [
					"Restaurant service"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "city",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "country",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "swLat",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "swLng",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "neLat",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "neLng",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "restaurant_type",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/SortRestaurantBy"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}